param: Add OFI_NCCL_PROGRESS to specify progress mode
This patch is to allow a parameter to specify the progress mode for the libfabric provider.  The provider's default progress mode is used; however, there is an issue with manual progress mode because NCCL stops polling the receiver's completion queue if there are no outstanding receive requests.  This is an issue if the sender's receive completion queue dropped the ACKs sent by the receiver and the receiver already posted the receive completion.  In this case the receiver is not polled on to determine that it needs to resend the ACKs causing the sender to be unable to determine the send completed.  There needs to be a way to override using the provider's default progress mode to force auto progress mode in order to get around the issue seen with manual progress mode.

Signed-off-by: Lindsay Reiser <lindsay.reiser@cornelisnetworks.com>